Author an equivalent of the Ruby service on top of the Custom service
This service tries to behave like the DotCloud Ruby service but comes
with some customisations compared to the regular Ruby service:
- The ability to install system packages (very helpful to debug stuff);
- The ability to choose a specific Ruby version with rvm (default is
  1.9.2p290):
- Updated versions of Passenger and Nginx;
- The ability to use the Passenger debug tools: passenger-status and
  passenger-memory-stats.

The builder script can be easily modified to install something else
instead of Nginx+Passenger to serve the Rack application.

This service supports the `approot' directive but when you use it pay
attention to adjust the path in `buildscript`.